requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.2
coverage==7.8.0
cryptography==44.0.2
cycler==0.12.1
defusedxml==0.8.0rc2
Django==5.2
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-rest-knox==5.0.2
django-templated-mail==1.1.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
fonttools==4.57.0
gunicorn==23.0.0
hiredis==3.1.0
idna==3.10
kiwisolver==1.4.8
kombu==5.5.3
matplotlib==3.10.1
mysqlclient==2.2.7
numpy==2.2.5
oauthlib==3.2.2
packaging==24.2
phonenumbers==9.0.5
phonenumberslite==8.13.55
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pydot==3.0.4
pydotplus==2.0.2
PyJWT==2.10.1
pyparsing==3.2.3
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
qrcode==8.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
